+++ allocators/src/unsigned_integer_array.c
@@ -2,10 +2,11 @@
 
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 // Return the Length of the Dynamic Unsigned Integer Array
-int array_length(unsigned int *array[]) {
-  int index = 0;
+int8_t array_length(uint8_t *array[]) {
+  int8_t index = 0;
 
   while(array[index] != NULL) {
     index = index + 1;
@@ -17,9 +18,9 @@ int array_length(unsigned int *array[]) {
 // Clear all Elements from the Dynamic Unsigned Integer Array
 // Assigns the Value of all Elements to NULL
 // Returns the Array
-unsigned int clear_array(unsigned int *array[]) {  
-  int index;
-  int length = array_length(array);
+uint8_t clear_array(uint8_t *array[]) {  
+  int8_t index;
+  int8_t length = array_length(array);
     
   for (index = 0; index < length; index = index + 1) {
     array[index] = NULL;
@@ -31,17 +32,17 @@ unsigned int clear_array(unsigned int *array[]) {
 // Create Dynamic Unsigned Integer Array
 // Returns the Array
 // Returns NULL if Memory Allocation Failed
-unsigned int create_array(unsigned int *array[], int capacity) {  
+uint8_t create_array(uint8_t *array[], int8_t capacity) {  
   if (capacity == 0) {
     capacity = 4;
-    array = malloc(sizeof(int[capacity]));
+    array = malloc(sizeof(int8_t[capacity]));
     
     if (array == NULL) {    
       printf(stderr, "Memory Allocation Error\n");
       return **array;
     }
   } else if (capacity != 0) {
-    array = malloc(sizeof(int[capacity]));
+    array = malloc(sizeof(int8_t[capacity]));
 
     if (array == NULL) {    
       printf(stderr, "Memory Allocation Error\n");
@@ -57,11 +58,11 @@ unsigned int create_array(unsigned int *array[], int capacity) {
 // Shifts all Elements after it, to the Right
 // Returns the Array
 // Returns NULL if Memory Allocation Failed
-unsigned int insert_element(unsigned int *array[], int element, int index) {
-  int address;
-  int length = array_length(array);
+uint8_t insert_element(uint8_t *array[], int8_t element, int8_t index) {
+  int8_t address;
+  int8_t length = array_length(array);
 
-  array = realloc(array,sizeof(int[length + 1]));
+  array = realloc(array,sizeof(int8_t[length + 1]));
 
   if (array == NULL) {    
       printf(stderr, "Memory Allocation Error\n");
@@ -80,10 +81,10 @@ unsigned int insert_element(unsigned int *array[], int element, int index) {
 // Append an Element to the End of the Dynamic Unsigned Integer Array
 // Returns the Array
 // Returns NULL if Memory Allocation Failed
-unsigned int push_element(unsigned int *array[], int element) {
-  int length = array_length(array);
+uint8_t push_element(uint8_t *array[], int8_t element) {
+  int8_t length = array_length(array);
 
-  array = realloc(array,sizeof(int[length + 1]));
+  array = realloc(array,sizeof(int8_t[length + 1]));
     
   if (array == NULL) {    
       printf(stderr, "Memory Allocation Error\n");
@@ -100,16 +101,16 @@ unsigned int push_element(unsigned int *array[], int element) {
 // Returns the Element
 // Shifts all Elements after it, to the Left
 // Returns NULL if Memory Allocation Failed
-unsigned int remove_element(unsigned int *array[], int index) {
-  int address;
-  unsigned int *element = array[index];
-  int length = array_length(array);    
+uint8_t remove_element(uint8_t *array[], int8_t index) {
+  int8_t address;
+  uint8_t *element = array[index];
+  int8_t length = array_length(array);    
   
   for (address = index; address < length; address = address + 1) {
     array[address] = array[address + 1];
   }
   
-  array = realloc(array,sizeof(int[length - 1]));
+  array = realloc(array,sizeof(int8_t[length - 1]));
 
   if (array == NULL) {
       printf(stderr, "Memory Allocation Error\n");
